+++ CHANGES.rst
@@ -5,7 +5,8 @@ Here you can see the full list of changes between each hmmlearn release.
 
 next
 ----
-
+- Introduce learning HMMs with Variational Inference.  Support
+  Gaussian and Categorical Emissions.
 - Removed the deprecated ``iter_from_X_lengths`` and ``log_mask_zero``;
   ``lengths`` arrays that do not sum up to the entire array length are no
   longer supported.